Make `get_supported_configs()` return `list[KeyConfig]`
Modify `get_supported_configs()` prototype to return a list of
`KeyConfig` instances, rather than a complete `ProviderConfig`.  This
avoids a bit of redundancy and a potential source of errors, given that
`namespace` is already exposed on the plugin instance.  By moving
the construction of final `ProviderConfig` to `variantlib`, we take
the responsibility for filing `ProviderConfig.namespace`.

While at it, simplify the branch for no configs -- we can simply return
an empty list rather than adding an additional branch for `None`.