Load plugins automatically when using PluginLoader API
It seems that the refactoring related to singleton removal resulted
in `load_plugins()` no longer being called automatically -- after all,
we are no longer constructing the class.  We didn't catch this in tests
since the plugin-related tests always called it explicitly, and API
tests were not written yet.  Just ensure that the plugins are loaded
when any of the `PluginLoader` API is used.