Load plugins automatically when using PluginLoader API (#26)
* Reduce mocked plugin fixture scope to test

* Add an autouse fixture to flush PluginLoader cache

Add a fixture that automatically flushes PluginLoader cache after each
test.  This should ensure that tests don't silently leak plugins,
and that we don't rely on plugins being loaded implicitly somewhere.

* Load plugins automatically when using PluginLoader API

It seems that the refactoring related to singleton removal resulted
in `load_plugins()` no longer being called automatically -- after all,
we are no longer constructing the class.  We didn't catch this in tests
since the plugin-related tests always called it explicitly, and API
tests were not written yet.  Just ensure that the plugins are loaded
when any of the `PluginLoader` API is used.

* Skip plugin discovery when already loaded