Initial sketch of get_build_setup() API (#29)
Initial sketch of API to get variables specific to build environment.
The plugin method takes a list of variant properties (variantlib filters
them, so that only properties specific to given plugin are passed)
and returns a `dict[str, list[str]]` of variables.  If two or more plugins
return the same key, the values are combined.