Make WHEEL_NAME_VALIDATION_REGEX verbose and improve it (#39)
* Verbosely describe WHEEL_NAME_VALIDATION_REGEX and remove .dist-info

Refactor `WHEEL_NAME_VALIDATION_REGEX` to verbosely describe what it
does and make it easier to comprehend.  While at it, remove the support
for `.dist-info` groups that was tangled into the regex that we only use
for wheels.

* Add a <base_wheel_name> group to WHEEL_NAME_VALIDATION_REGEX

Add a `base_wheel_name` group to `WHEEL_NAME_VALIDATION_REGEX` to be
able to conveniently grab the base name without variant hash without
having to reconstruct it manually.