Support for evaluating "variants" environment specifiers (#32)
* Add support for evaluating variant expressions in requirements

Add a helper routine that can be used to process requirements
and evaluate the "variants"-based marker strings.  The use case is for
tools such as `build` that pass `build-system.requires` to the installer
(pip or uv), to be able to filter the requirements based on selected
variants.

* Use regex to process environment markers instead

Use regex to preprocess "variants" environment markers rather than
the internal API of packaging.  While at it, add support for negative
"not in" expressions, and effectively fix support for dealing with
compound statements.

Rather than attempting to implement complete boolean logic to process
various conditions, we replace the variant markers with a "true"
and "false" conditions that pip/uv can handle.

* Split markers without packaging

The requirement syntax is simple enough to make it possible to trivially
split marker part without having to use `packaging`.  This way, we don't
have to fork it immediately to add support for `variants` environment
markers, that otherwise cause parsing to fail.

* Combine invalid envspec checks into a single regexp

* Detect != operator too

* Allow empty list of VariantProperty to envspec processing

* Document InvalidVariantEnvSpecError