Use new sorting API in `variantlib.api` (#44)
* Remove namespace_priority_dict from get_variant_hashes_by_priority()

Remove the `namespace_priority_dict` that is not used by consumers,
and is not going to be the part of the final API, before starting
to rewrite the function.

* Move get_combinations() helper into tests

* Add to_list_of_properties() method to ProviderConfig

Add a method to return a flattened list of properties
for a `ProviderConfig`.

* Replace combination tests with more salient example

* Switch to the new sorting logic

* Invent fake names for tests

* Update test_combinations to use test plugins

* Remove `variantlib.combination`

Remove the `variantlib.combination` module, moving the remaining code
directly into `variantlib.api` and the tests to `test_api`.