Add newlines to variant hash data to ensure uniqueness
Add newlines to variant property strings when computing the variant hash
in order to ensure that different sets of variant properties always
result in different hashes.

For example, previously the variant:

    a :: b :: cx
    d :: e :: f

would result in the same hash as:

    a :: b :: c
    xd :: e :: f

Appending a newline to every entry seems the most logical solution,
as it effectively means that variant properties are serialized
into a text-friendly, newline-separated list of strings.