Reintroduce entry point plugin loading for variantlib commands
Replace the `CLIPluginLoader` class with `EntryPointPluginLoader` that
loads all plugins found in installed entry points.  Use it in most
of the variantlib commands to make them much more convenient to use.

To simplify the test suite, switch the main fixture to use
`ManualPluginLoader`, and instead add a single test for
`EntryPointPluginLoader` specifically.