Restore entry points based loader, and use it for CLI commands (#90)
* Reintroduce entry point plugin loading for variantlib commands

Replace the `CLIPluginLoader` class with `EntryPointPluginLoader` that
loads all plugins found in installed entry points.  Use it in most
of the variantlib commands to make them much more convenient to use.

To simplify the test suite, switch the main fixture to use
`ManualPluginLoader`, and instead add a single test for
`EntryPointPluginLoader` specifically.

* Support context manager API to `ManualPluginLoader`

Add a support for context manager API to `ManualPluginLoader` for
consistency with other context manager classes.  It is not really
necessary, but having consistent API will make use easier.

* make-variant: Automatically determine plugin-api from entry points

* Simplify `mocked_plugin_loader` fixture

* update-pyproject-toml: New command to add/remove providers

* make-variant: Read variant metadata from `pyproject.toml`

---------

Co-authored-by: Michał Górny <mgorny@quansight.com>