Make feature and property priorities a dict (#96)
* Remove `config setup` command

The command will be too hard to update for the upcoming priority
changes, and it has little usefulness now that priorities are configured
by packages.

* Change feature priorities to a per-namespace dict

This is the first part of:
https://github.com/wheelnext/pep_xxx_wheel_variants/issues/21#issuecomment-2894583251

It replaces the current flat feature priority list with a dictionary
that adjusts priorities per namespace.

* Change property priorities into per-feature dict

Change property priorities to reorder only within the feature,
and update the sort algorithms to work within the new framework.

* Remove `get_variant_property_priorities_tuple`

It does not make sense anymore with the hierarchical sorting.

* Update `variants.dist.toml`

* Fix `config show` command