Support for PEP517-style module plugins and optional `plugin-api` (#97)
* Support non-callable `plugin-api`

Support using a `plugin-api` valuable that is not callable, and use
it as-is.  Currently, this is exemplified by an already instantiated
object, but the goal is to support a whole module, like PEP 517 does.

* Support specifying a module for `plugin-api`

* Convert installable test plugin to a module-level plugin-api

* Make `plugin-api` optional

Make the `plugin-api` key optional, defaulting to a value inferred
from `requires` when not specified.  This should make plugins easier
to use in the most common cases, while providing the usual override
for specific needs (and backwards compatibility).

https://github.com/wheelnext/pep_xxx_wheel_variants/issues/38#issuecomment-2894441630