Introduce a `WheelMetadata` subclass
Introduce a `WheelMetadata` subclass that provides a more convenient
API to access the `variant.json` data from wheels.  Notably, it adds
verification that the file contains only one variant (with an option
to verify its hash against the filename) and properties to query
the variant description and hash, and to set it.