Initial support for `optional-providers` key
Add initial support for `optional-providers` key that is like
`providers`, except that the providers specified there are optional
and therefore are not installed by default.  This aims to provide
a cleaner interface to making some plugins opt-in than the previous
idea of controlling them via `default-proirities.namespaces`.

Currently, just the baseline code without an actual API to enable
installing them.