Introduce support for dynamic config return values in plugins (#106)
* Introduce support for dynamic config return values in plugins

Add support for the plugins dynamically adjusting the return values
for `get_all_configs()` and `get_supported_configs()` based on input
`known_properties` values.  The values are either provided based
on the variants being validated (for `get_all_configs()`) or based
on all properties found in available wheels (for
`get_supported_configs()`).  The dynamic API is entirely optional --
the plugin can simply ignore the additional argument and return a fixed
list as usual.  It also permits for hybrid plugins that combine static
and dynamic properties.

* Remove the default argument from protocol

* Add a `dynamic` property to distinguish dynamic plugins

* Use `frozenset` in plugin API

* Use a stricter value regex

* Extend failing regex tests with \t