Introduce a dedicated `validate_properties()` method in plugin API (#107)
* Introduce a dedicated `validate_properties()` method in plugin API

* Merge the obsolete `BasePluginLoader._get_configs()` helper

* Change the API to a single-property `validate_property()`

Given that the benefit (if any) from validating all properties at once
is minimal, change the protocol to much simpler `validate_property()`
method.  This avoids problems with custom property types that aren't
suitable as `dict` keys, as well as the more general problem of "what
if dict contains different keys than the properties we've given it".