Null variant labels (#110)
* Add a `get_variant_label()` helper

* Use 'null' label for the null variant

* Use `sha256("")` for null variant hash

Since we are no longer using it as a label, we do not need
a recognizable special value here.  Instead, just use hash of an empty
string, so we can remove the special casing in code.

* Restore explicit checks for `--variant-label` use

* Remove `NULL_VARIANT_HASH`

Remove `NULL_VARIANT_HASH` since it resembles magic constant too much.
In the few places where we need it, either call hash directly, or get
it from VariantDescription.

Signed-off-by: Michał Górny <mgorny@quansight.com>